Inital Commit for Selenic.API in .NET 5
This project is intended for personal use only. It is a rewrite of an
earlier written application (NodeJS) for one of the projects at the
Avans University of Applied Sciences.

Main goal of this project is to get familiar with .NET 5 and their
frameworks.

For now I'm intending to use or include:
- Serilog (Semantic Logging), without sending the logs to a centralized
  location. (Which is fairly easy, just include and configure an extra
  sink)
- Background threads/workers for scheduling expensive tasks to happen
  background or to execute scheduled events
- Some sort of authentication
- Something Else?

All rights reserved.